+++ client/src/features/relatedLists/ListItem.tsx
@@ -12,6 +12,7 @@ import DescriptionDetail from '../../styles/shared/DescriptionDetail'
 import theme from '../../styles/theme'
 import RecordLink from '../common/RecordLink'
 import { formatFacetedSearchJson } from '../../lib/parse/search/halLinkHelper'
+import { pushSiteImproveEvent } from '../../lib/siteImprove'
 
 interface IProps {
   uri: string
@@ -50,6 +51,8 @@ const ListItem: React.FC<IProps> = ({
   }
 
   const searchQ = formatFacetedSearchJson(criteria, searchTerm, uri)
+  const linkLabel = `Show all ${count} result${count !== 1 ? 's' : ''}`
+
   return (
     <StyledRow
       key={uri}
@@ -76,10 +79,16 @@ const ListItem: React.FC<IProps> = ({
               prevPath: pathname,
               targetName: `/view/results/${tab}?q=${searchQ}&openSearch=false`,
             }}
+            onClick={() =>
+              pushSiteImproveEvent(
+                `${uri} Faceted Results Search Link`,
+                'Selected',
+                linkLabel,
+              )
+            }
             data-testid={`list-item-link-${index}`}
           >
-            Show all {count} result
-            {count !== 1 && `s`}
+            {linkLabel}
           </Link>
         </DescriptionDetail>
       </Col>