+++ client/src/features/relatedLists/SearchResultRelatedList.tsx
@@ -10,6 +10,7 @@ import { getEstimates } from '../../lib/parse/search/searchResultParser'
 import RecordLink from '../common/RecordLink'
 import { searchScope } from '../../config/searchTypes'
 import { getAllParamsFromHalLink } from '../../lib/parse/search/halLinkHelper'
+import { pushSiteImproveEvent } from '../../lib/siteImprove'
 
 interface IProps {
   url: string
@@ -59,6 +60,7 @@ const SearchResultRelatedList: React.FC<IProps> = ({ url, scope, data }) => {
   const searchString = `${searchQ}&openSearch=false${
     sort !== null ? `&${resultsEndpoint[0]}s=${sort}` : ''
   }`
+  const linkLabel = `Show all ${estimate} result${estimate !== 1 ? 's' : ''}`
 
   return (
     <React.Fragment>
@@ -74,10 +76,12 @@ const SearchResultRelatedList: React.FC<IProps> = ({ url, scope, data }) => {
               prevPath: pathname,
               targetName: `/view/results/${newScope}${searchString}`,
             }}
+            onClick={() =>
+              pushSiteImproveEvent('Search Results Link', 'Selected', linkLabel)
+            }
             data-testid="search-related-list-link"
           >
-            Show all {estimate} result
-            {estimate !== 1 && `s`}
+            {linkLabel}
           </Link>
         </div>
       </StyledSearchLink>